♻️ refactor(env): relocate RedisConfig type to @lobechat/types
