+++ packages/types/src/index.ts
@@ -28,6 +28,7 @@ export * from './message';
 export * from './meta';
 export * from './plugins';
 export * from './rag';
+export * from './redis';
 export * from './search';
 export * from './serverConfig';
 export * from './service';
