+++ packages/types/src/redis.ts
@@ -1,0 +1,9 @@
+export type RedisConfig = {
+  database?: number;
+  enabled: boolean;
+  password?: string;
+  prefix: string;
+  tls: boolean;
+  url: string;
+  username?: string;
+};
