+++ src/envs/redis.ts
@@ -1,7 +1,7 @@
 import { createEnv } from '@t3-oss/env-core';
 import { z } from 'zod';
 
-import { type RedisConfig } from '@/libs/redis';
+import { type RedisConfig } from '@lobechat/types';
 
 const parseNumber = (value?: string) => {
   const parsed = Number.parseInt(value ?? '', 10);
